% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymarkdown.R
\name{get_html}
\alias{get_html}
\title{Get the editor's HTML output}
\usage{
get_html()
}
\value{
A string containing the HTML code created from text entered into the Markdown editor.
}
\description{
Get the editor's HTML output
}
