% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymarkdown.R
\name{init_editor}
\alias{init_editor}
\title{Initialize the editor}
\usage{
init_editor(
  min_height,
  height,
  preview_style,
  preview_highlight,
  initial_edit_type,
  hide_mode_switch,
  language,
  initial_value
)
}
\arguments{
\item{min_height}{The editor's miniminum height value (pixels). Default = '300px'. Overwritten by the height parameter.}

\item{height}{The editor's height value, applied as a border-box. Example values include "300px", "100\%", and "auto". Defaults to "500px".}

\item{preview_style}{The Markdown editor's preview style. Either "tab" or "vertical". Default is "tab".}

\item{preview_highlight}{Should the Markdown Editor's HTML preview have a highlighted element corresponding to the cursor position in the Markdown editor. Default is FALSE. be highligh}

\item{initial_edit_type}{Initial editor type: "markdown" or "wysiwyg". Default is "markdown".}

\item{hide_mode_switch}{Should the user be able to switch the editor mode from "wysiwyg" to "markdown" or vice versa? Default is TRUE.}

\item{language}{Editor language ISO code. Defaults to "en-us".}

\item{initial_value}{Should the editor have text already present? If so, supply a character string. Default is NULL.}
}
\value{
A script defining how to initialize the editor based on user specifications.
}
\description{
Initialize the editor
}
\keyword{internal}
