% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_editor.R
\name{r}
\alias{r}
\title{Show the editor}
\usage{
r(.id)
}
\arguments{
\item{.id}{The \code{inputId} of the editor.}
}
\description{
Show the editor
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinymarkdown)
  ui <- fluidPage(
    mdInput(inputId = "myEditor", height = "500px", hide_mode_switch = F)
  )

  server <- function(input, output, session) {
    # Hide the editor
    observeEvent(input$hide, {
      hide_ediotr(.id = "myEditor")
    })

    # Show the editor
    obseveEvent(input$show, {
      show_editor(.id = "myEditor")
    })
  }

  shinyApp(ui, server)
}

}
