% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_input_ids.R
\name{sanitize_input_ids}
\alias{sanitize_input_ids}
\title{Sanitize Input IDs}
\usage{
sanitize_input_ids(inputId)
}
\arguments{
\item{inputId}{The input ID to sanitize.}
}
\description{
\code{{shinymarkdown}} doesn't work if the inputId contains a "-" in its name. And "-" is
particularly used by Shiny namespaces, so it's important to perform these ID fixes.
}
