% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useShinyMarkdownEditor.R
\name{useShinyMarkdownEditor}
\alias{useShinyMarkdownEditor}
\title{Set up a Shiny app to use shinymarkdowneditor}
\usage{
useShinyMarkdownEditor()
}
\description{
This function must be called from a Shiny app's UI in order for all other
shinymarkdowneditor functions to work.
You can call \code{useShinyMarkdownEditor()} from anywhere inside the UI, as long as the final
app UI contains the result of \code{useShinyMarkdownEditor()}.
}
