% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownEditorInput.R
\name{markdownEditorInput}
\alias{markdownEditorInput}
\title{Create a Markdown Editor input control}
\usage{
markdownEditorInput(inputId)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}
}
\description{
Create a Markdown Editor input control
}
